requirements.txt
aiofiles==22.1.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.4.0
blinker==1.9.0
browser-cookie3==0.19.1
build==1.2.2.post1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
f2==0.0.1.6
Flask==3.1.0
gmssl==3.2.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
importlib-resources==6.1.0
iniconfig==2.0.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpath-ng==1.6.0
lz4==4.3.3
m3u8==3.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
pip-tools==7.4.1
pluggy==1.5.0
ply==3.11
protobuf==4.23.0
pycryptodomex==3.21.0
pydantic==2.6.4
pydantic_core==2.16.3
PyExecJS==1.5.1
Pygments==2.18.0
pypng==0.20220715.0
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pytest==8.2.1
pytest-asyncio==0.21.1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.32.3
rich==13.7.1
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.2
websockets==14.1
Werkzeug==3.1.3
wheel==0.45.1
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0